DOSIS-999 Rebased and brought in logger
