+++ tests/service_automation/tests/config_template.yaml
@@ -1,0 +1,23 @@
+environment: "dev"
+
+environments:
+  dev:
+    bucket_type:
+    project:
+    base_url:
+    api_key:
+    search_engine_id:
+
+  nonprod:
+    bucket_type:
+    project:
+    base_url:
+    api_key:
+    search_engine_id:
+
+  prod:
+    bucket_type:
+    project:
+    base_url:
+    api_key:
+    search_engine_id:
