+++ tests/service_automation/tests/step_definitions/is_api_steps/test_api_google_search.py
@@ -8,14 +8,6 @@ from config import config  # Ensure Config is correctly imported
 # Load feature file
 scenarios("./is_api_features/test_api_google_search.feature")
 
-# @pytest.fixture(scope="session")
-# def api_request_context():
-#     """Initialize Playwright APIRequestContext"""
-#     with sync_playwright() as p:
-#         request_context = p.request.new_context()
-#         yield request_context
-#         request_context.dispose()
-
 @pytest.fixture
 def api_response():
     """Fixture to store API response for logging in reports."""
